requirements.txt
Pillow~=8.0.1
numpy~=1.19.4
pip~=20.3.3
wheel~=0.36.2
rsa~=4.6
pyasn1~=0.4.8
six~=1.15.0
gast~=0.3.3
h5py~=2.10.0
setuptools~=51.1.0.post20201221
scipy~=1.5.4
matplotlib~=3.3.3
cachetools~=4.2.0
certifi~=2020.12.5
urllib3~=1.26.2
requests~=2.25.1
joblib~=1.0.0
chardet~=4.0.0
scikit-learn~=0.23.2
threadpoolctl~=2.1.0
python-dateutil~=2.8.1
Markdown~=3.3.3
oauthlib~=3.1.0
idna~=2.10
Werkzeug~=1.0.1
cycler~=0.10.0
pyparsing~=2.4.7
kiwisolver~=1.3.1
tensorflow~=2.4.0
tensorboard~=2.4.0
flatbuffers~=1.12
wrapt~=1.12.1
astunparse~=1.6.3
termcolor~=1.1.0
pytesseract~=0.3.7
zipp~=3.4.0